# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: beancount-gs API
  version: 2.0.0
  description: |
    beancount-gs V2 REST API。命名遵循 beancount 领域术语（narration / postings / units / cost /
    operating_currency / opening_balances 等），JSON 字段统一 snake_case。

    认证：
    - cookieAuth：浏览器会话（GitHub OAuth2 登录后下发 httpOnly Cookie）。
    - bearerAuth：Agent / 第三方通过 API Key 接入（MCP 亦使用同一套 Key）。

    并发：所有写接口要求请求头 `If-Revision-Match: <ledger_revision>`（或用 body 字段
    `base_revision`），不匹配返回 409 LEDGER_STALE 并携带当前修订号。
servers:
  - url: http://localhost:10000/api/v2
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /auth/github/callback:
    get:
      tags: [auth]
      summary: GitHub OAuth 回调
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        "302": { description: 成功后重定向到前端，并下发会话 Cookie }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /teams:
    get:
      tags: [teams]
      summary: 当前用户所在工作区列表
      responses:
        "200":
          description: 工作区列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Team" }
    post:
      tags: [teams]
      summary: 新建工作区（创建者成为 owner）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamCreate" }
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /ledgers:
    get:
      tags: [ledgers]
      summary: 当前用户可访问的账本列表
      responses:
        "200":
          description: 账本列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Ledger" }
    post:
      tags: [ledgers]
      summary: 新建账本（按模板初始化 bean 文件结构）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LedgerCreate" }
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ledger" }
  /ledgers/{ledger_id}:
    get:
      tags: [ledgers]
      summary: 账本详情（含修订号与我的角色）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      responses:
        "200":
          description: 账本详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ledger" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [ledgers]
      summary: 删除账本（owner，二次确认参数 confirm=true）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - name: confirm
          in: query
          required: true
          schema: { type: boolean }
      responses:
        "204": { description: 已删除 }
        "403": { $ref: "#/components/responses/Forbidden" }
  /ledgers/import:
    post:
      tags: [ledgers]
      summary: 从备份 zip 导入并新建账本
      description: 解压校验（目录结构 + bean-check）后创建新账本；zip 必须包含根目录 index.bean。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [team_id, name, file]
              properties:
                team_id: { type: string }
                name: { type: string, description: 新账本名称 }
                operating_currency: { type: string, default: CNY }
                skip_validation:
                  type: boolean
                  default: false
                  description: 跳过 bean-check 语法校验（用于迁移非标准语法的历史账本）
                file: { type: string, format: binary, description: 账本备份 zip }
      responses:
        "201":
          description: 导入并创建成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BackupImportResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /ledgers/{ledger_id}/import:
    post:
      tags: [ledgers]
      summary: 导入备份 zip 到已有账本
      description: 覆盖账本文件（被覆盖文件先快照到 bak/），通过 bean-check 校验后写入并递增修订号。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                skip_validation:
                  type: boolean
                  default: false
                  description: 跳过 bean-check 语法校验（用于迁移非标准语法的历史账本）
                file: { type: string, format: binary, description: 账本备份 zip }
      responses:
        "200":
          description: 导入成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BackupImportResult" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /ledgers/{ledger_id}/revision:
    get:
      tags: [ledgers]
      summary: 当前修订号（写操作前检查）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      responses:
        "200":
          description: 修订号
          content:
            application/json:
              schema:
                type: object
                required: [ledger_id, revision]
                properties:
                  ledger_id: { type: string }
                  revision: { type: integer, example: 128 }
  /ledgers/{ledger_id}/months:
    get:
      tags: [ledgers]
      summary: 有交易的月份列表（最近在前）
      description: 供前端月份筛选下拉使用，数据来自账本实际交易（distinct year/month）。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      responses:
        "200":
          description: 月份列表，如 ["2026-08","2026-07"]
          content:
            application/json:
              schema:
                type: array
                items: { type: string, example: "2026-08" }
  /ledgers/{ledger_id}/accounts:
    get:
      tags: [accounts]
      summary: 账户列表（含余额 / 持仓）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - name: status
          in: query
          schema: { type: string, enum: [open, closed], default: open }
      responses:
        "200":
          description: 账户列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Account" }
    post:
      tags: [accounts]
      summary: 开户（写入 account/*.bean 的 open 指令）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountOpen" }
      responses:
        "201":
          description: 开户成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "409": { $ref: "#/components/responses/Conflict" }
  /ledgers/{ledger_id}/accounts/batch:
    post:
      tags: [accounts]
      summary: 批量开户（一次写入多个 open 指令）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountOpenBatch" }
      responses:
        "201":
          description: 批量开户结果（含跳过项）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountOpenBatchResult" }
        "409": { $ref: "#/components/responses/Conflict" }
  /ledgers/{ledger_id}/accounts/{account}:
    get:
      tags: [accounts]
      summary: 账户详情（持仓 / 盈亏 / 余额趋势）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/AccountParam" }
      responses:
        "200":
          description: 账户详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountDetail" }
    post:
      tags: [accounts]
      summary: 关闭账户（close 指令）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/AccountParam" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [closed_on]
              properties:
                closed_on: { type: string, format: date }
      responses:
        "200":
          description: 已关闭
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
    put:
      tags: [accounts]
      summary: 重新开户（close 后追加 open 指令恢复使用）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/AccountParam" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opened_on]
              properties:
                opened_on: { type: string, format: date }
      responses:
        "200":
          description: 已重新开户
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "409": { $ref: "#/components/responses/Conflict" }
  /ledgers/{ledger_id}/accounts/{account}/balance:
    post:
      tags: [accounts]
      summary: 期初对账（写入 pad + balance 指令）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/AccountParam" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, number]
              properties:
                date: { type: string, format: date }
                number: { type: string, description: decimal 字符串 }
      responses:
        "200": { description: 对账指令已写入 }
  /ledgers/{ledger_id}/account-types:
    get:
      tags: [accounts]
      summary: 账户类型映射（前缀 → 名称）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      responses:
        "200":
          description: 类型映射
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AccountTypeMapping" }
    post:
      tags: [accounts]
      summary: 新增账户类型
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountTypeMapping" }
      responses:
        "201": { description: 已新增 }
  /ledgers/{ledger_id}/transactions:
    get:
      tags: [transactions]
      summary: 交易列表
      description: 按条件筛选交易，默认按日期倒序。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - name: from
          in: query
          schema: { type: string, format: date }
          description: 起始日期
        - name: to
          in: query
          schema: { type: string, format: date }
          description: 结束日期
        - name: month
          in: query
          schema: { type: string, pattern: "^\\d{4}-\\d{2}$" }
          description: 月份，如 2026-08
        - name: account
          in: query
          schema: { type: string }
          description: 账户精确匹配
        - name: account_type
          in: query
          schema: { type: string, enum: [Assets, Liabilities, Income, Expenses, Equity] }
          description: 账户根类型前缀匹配
        - name: q
          in: query
          schema: { type: string }
          description: 收款方 / 描述模糊搜索
        - name: tag
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 10000, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: 交易列表
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Transaction" }
                  total: { type: integer }
    post:
      tags: [transactions]
      summary: 新建交易
      description: 服务端校验借贷平衡；外币分录自动写入 price 指令并生成 cost。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionCreate" }
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /ledgers/{ledger_id}/transactions/{transaction_id}:
    get:
      tags: [transactions]
      summary: 交易详情
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/TransactionId" }
      responses:
        "200":
          description: 交易详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
    put:
      tags: [transactions]
      summary: 更新交易（整体替换）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/TransactionId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [transactions]
      summary: 删除交易
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/TransactionId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
      responses:
        "204": { description: 已删除 }
  /ledgers/{ledger_id}/transactions/{transaction_id}/raw:
    get:
      tags: [transactions]
      summary: 交易原始 bean 文本
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/TransactionId" }
      responses:
        "200":
          description: 原始文本
          content:
            text/plain: { schema: { type: string } }
    put:
      tags: [transactions]
      summary: 用原始文本更新交易
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/TransactionId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
      requestBody:
        required: true
        content:
          text/plain: { schema: { type: string } }
      responses:
        "200": { description: 已更新 }
...
  /ledgers/{ledger_id}/stats/total:
    get:
      tags: [stats]
      summary: 按账户类型汇总
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/Month" }
        - name: account
          in: query
          schema: { type: string }
          description: 账户前缀过滤，如 Expenses
      responses:
        "200":
          description: 类型 → 金额
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
  /ledgers/{ledger_id}/stats/payee:
    get:
      tags: [stats]
      summary: 收款方统计（金额 / 笔数 / 均值）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/Month" }
        - name: account
          in: query
          schema: { type: string }
          description: 账户前缀过滤，如 Expenses
        - name: type
          in: query
          schema: { type: string, enum: [total, count, avg] }
      responses:
        "200":
          description: 收款方统计列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StatsPayee" }
  /ledgers/{ledger_id}/stats/account-trend:
    get:
      tags: [stats]
      summary: 账户余额趋势
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/Month" }
        - name: account
          in: query
          schema: { type: string }
          description: 账户前缀过滤，如 Assets
        - name: type
          in: query
          schema: { type: string, enum: [day, month, year, sum] }
      responses:
        "200":
          description: 趋势点列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StatsPoint" }
  /ledgers/{ledger_id}/stats/account-flow:
    get:
      tags: [stats]
      summary: 账户资金流向（桑基图数据）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/Month" }
        - name: account
          in: query
          schema: { type: string }
          description: 账户前缀过滤，如 Expenses
      responses:
        "200":
          description: 节点与边
          content:
            application/json:
              schema:
                type: object
                required: [nodes, links]
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                      properties: { name: { type: string } }
                  links:
                    type: array
                    items:
                      type: object
                      properties:
                        source: { type: integer }
                        target: { type: integer }
                        value: { type: string }
...
  /ledgers/{ledger_id}/imports/{source}:
    post:
      tags: [imports]
      summary: 上传账单 CSV 并预览（不落账）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - name: source
          in: path
          required: true
          schema: { type: string, enum: [alipay, wechat, icbc, abc] }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: 预览结果（含 AI 分类建议）
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ImportRow" }
  /ledgers/{ledger_id}/imports/{source}/confirm:
    post:
      tags: [imports]
      summary: 确认导入（写入账本）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
        - name: source
          in: path
          required: true
          schema: { type: string, enum: [alipay, wechat, icbc, abc] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows]
              properties:
                rows:
                  type: array
                  items: { $ref: "#/components/schemas/ImportRow" }
      responses:
        "201":
          description: 导入结果（成功 / 失败列表）
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: integer }
                  failed:
                    type: array
                    items:
                      type: object
                      properties:
                        index: { type: integer }
                        reason: { type: string }
  /ledgers/{ledger_id}/currencies:
    get:
      tags: [currencies]
      summary: 币种与最新汇率
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      responses:
        "200":
          description: 币种列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Currency" }
    post:
      tags: [currencies]
      summary: 新增币种
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currency, name]
              properties:
                currency: { type: string, example: USD }
                name: { type: string, example: 美元 }
                symbol: { type: string, example: "$" }
      responses:
        "201": { description: 已新增 }
  /ledgers/{ledger_id}/currencies/sync:
    post:
      tags: [currencies]
      summary: 同步最新汇率（写入 price 指令）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/IfRevisionMatch" }
      responses:
        "200":
          description: 同步后的币种列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Currency" }
  /ledgers/{ledger_id}/ai/record:
    post:
      tags: [ai]
      summary: 自然语言记账（生成待确认交易草稿）
      description: AI 不直接写账；返回草稿与建议，由调用方确认后调用交易创建接口。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, example: "昨天星巴克咖啡 38 元" }
      responses:
        "200":
          description: 待确认交易草稿
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiRecordResult" }
  /ledgers/{ledger_id}/ai/record/batch:
    post:
      tags: [ai]
      summary: 自然语言批量记账（一次生成多条待确认交易草稿）
      description: AI 不直接写账；一次可描述多笔交易，返回草稿数组，由调用方确认后逐笔调用创建接口。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, example: "昨天星巴克 38 元\n前天打车 120 元" }
      responses:
        "200":
          description: 待确认交易草稿数组
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiRecordBatchResult" }
  /ledgers/{ledger_id}/ai/record/chat:
    post:
      tags: [ai]
      summary: 对话式批量记账（多轮生成 / 调整草稿）
      description: 传入对话历史与当前草稿，AI 返回调整后的完整草稿数组；前端勾选后逐笔创建。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiRecordChatRequest" }
      responses:
        "200":
          description: 待确认交易草稿数组
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiRecordBatchResult" }
  /ledgers/{ledger_id}/ai/summarize:
    post:
      tags: [ai]
      summary: 生成账本总结
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/Month" }
      responses:
        "200":
          description: 文字总结
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary: { type: string }
  /ledgers/{ledger_id}/ai/insights:
    get:
      tags: [ai]
      summary: 洞察与异常检测（重复扣款 / 大额支出 / 环比异常）
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
        - { $ref: "#/components/parameters/Month" }
      responses:
        "200":
          description: 洞察列表
          content:
            application/json:
              schema:
                type: object
                required: [insights]
                properties:
                  insights:
                    type: array
                    items:
                      type: object
                      required: [type, message]
                      properties:
                        type: { type: string, enum: [duplicate, large_expense, trend_change] }
                        message: { type: string }
  /ledgers/{ledger_id}/ai/accounts:
    post:
      tags: [ai]
      summary: AI 批量设计账户（生成待确认账户列表）
      description: AI 不直接写账；返回建议账户列表，由调用方确认后调用批量开户接口。
      parameters:
        - { $ref: "#/components/parameters/LedgerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, example: "我需要这些账户：招商银行储蓄卡、日常餐饮、工资收入" }
      responses:
        "200":
          description: 建议账户列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiAccountsResult" }
...
  /api-keys/{key_id}:
    delete:
      tags: [integrations]
      summary: 吊销 API Key
      parameters:
        - name: key_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: 已吊销 }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: GitHub OAuth2 登录后下发的 httpOnly 会话 Cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: bgsk_
      description: API Key（MCP 与 REST 通用）
...
  responses:
...
    Unauthorized:
      description: 未登录或会话失效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unprocessable:
      description: 语义校验失败（借贷不平衡 / bean-check 失败 / 路径越界）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Team:
      type: object
      required: [id, name, role]
      properties:
        id: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        member_count: { type: integer }
        ledger_count: { type: integer }
        created_at: { type: string, format: date-time }
    TeamCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, example: 家庭 }
...
    Role:
      type: string
      enum: [owner, editor, viewer]
...